+++ src/app/shared/ase-loader/ase-loader.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'ase-loader',
+  standalone: true,
+  imports: [],
+  templateUrl: './ase-loader.component.html',
+  styleUrl: './ase-loader.component.scss'
+})
+export class AseLoaderComponent {
+
+}