+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 - Add Cart icon component
 - Add Cart items and quantity service
 - Add amiibo list filters
+- Add loader for home and amiibo detail screens